requirements.txt
# Django
Django==3.2.4

# Django REST Framework
djangorestframework==3.12.4

# Python decouple
python-decouple==3.4